Add Init command (#4)
* Add support for middleware before all commands

The middleware will replace each command modules handler with the
middleware.  It will handle setting the logger settings, and loading
the config file.  As well as handle the exiting of the commands.

Still need to bring "generate" up to speed

* Change the naming of the config options

* Finish adding middleware to handle loading of config

Update generate handler to use new middleware.
Update most log statements to use the chalk template literal instead
of the function calls

* Move some auth functions out of genereate so they can be used by init

* Implement a custom version of inquirer's prompt

For easily adding a suffix to all the prompts